+++ Arex388.OpenWeatherMap/OpenWeatherMapClient.cs
@@ -13,45 +13,88 @@ internal sealed class OpenWeatherMapClient(
 	IOpenWeatherMapClient {
 	private static readonly JsonSerializerOptions _jsonSerializerOptions = new() {
 		Converters = {
-			new CurrentWeatherResponseJsonConverter(),
-			new WeatherJsonConverter()
+			new HistoricalWeatherResponseJsonConverter(),
+			new LegacyCurrentWeatherResponseJsonConverter(),
+			new LegacyWeatherJsonConverter(),
+			new WeatherJsonConverter(),
+			new WeatherConditionJsonConverter()
 		},
 		PropertyNamingPolicy = JsonNamingPolicy.SnakeCaseLower
 	};
 
-	private readonly IValidator<CurrentWeather.Request> _currentWeatherValidator = services.GetRequiredService<IValidator<CurrentWeather.Request>>();
+	private readonly IValidator<HistoricalWeather.Request> _historicalWeatherValidator = services.GetRequiredService<IValidator<HistoricalWeather.Request>>();
+	private readonly IValidator<LegacyCurrentWeather.Request> _legacyCurrentWeatherValidator = services.GetRequiredService<IValidator<LegacyCurrentWeather.Request>>();
+	//private readonly IValidator<HistoricalWeather.Request> _historicalWeatherValidator = services.GetRequiredService<IValidator<HistoricalWeather.Request>>();
 	private readonly HttpClient _httpClient = httpClient ?? services.GetRequiredService<IHttpClientFactory>().CreateClient(nameof(IOpenWeatherMapClient));
 	private readonly OpenWeatherMapClientOptions _options = options ?? services.GetRequiredService<OpenWeatherMapClientOptions>();
 
 	public Guid Id { get; } = Guid.NewGuid();
 
-	public Task<CurrentWeather.Response> CurrentWeatherAsync(
+	public Task<HistoricalWeather.Response> HistoricalWeatherAsync(
 		decimal latitude,
 		decimal longitude,
-		Units units,
-		CancellationToken cancellationToken = default) => CurrentWeatherAsync(new CurrentWeather.Request {
+		DateTimeOffset timestampAt,
+		CancellationToken cancellationToken = default) => HistoricalWeatherAsync(new HistoricalWeather.Request {
 			Latitude = latitude,
 			Longitude = longitude,
-			Units = units
+			TimestampAt = timestampAt,
+			Units = _options.Units
 		}, cancellationToken);
 
-	public async Task<CurrentWeather.Response> CurrentWeatherAsync(
-		CurrentWeather.Request request,
+	public async Task<HistoricalWeather.Response> HistoricalWeatherAsync(
+		HistoricalWeather.Request request,
 		CancellationToken cancellationToken = default) {
 		if (cancellationToken.IsSupportedAndCancelled()) {
-			return CurrentWeather.Response.Cancelled;
+			return HistoricalWeather.Response.Cancelled;
 		}
 
 		// ReSharper disable once MethodHasAsyncOverloadWithCancellation
-		var validationResult = _currentWeatherValidator.Validate(request);
+		var validationResult = _historicalWeatherValidator.Validate(request);
 
 		if (!validationResult.IsValid) {
-			return CurrentWeather.Response.Invalid(validationResult);
+			return HistoricalWeather.Response.Invalid(validationResult);
 		}
 
 		try {
 			var response = await _httpClient.GetAsync(request.GetEndpoint(_options), cancellationToken).ConfigureAwait(false);
-			var currentWeather = await response.Content.ReadFromJsonAsync<CurrentWeather.Response>(_jsonSerializerOptions, cancellationToken).ConfigureAwait(false);
+			var historicalWeather = await response.Content.ReadFromJsonAsync<HistoricalWeather.Response>(_jsonSerializerOptions, cancellationToken).ConfigureAwait(false);
+
+			historicalWeather!.Errors = historicalWeather.Error.HasValue()
+				? [historicalWeather.Error!]
+				: historicalWeather.Errors;
+
+			return historicalWeather;
+		} catch {
+			return HistoricalWeather.Response.Failed;
+		}
+	}
+
+	public Task<LegacyCurrentWeather.Response> LegacyCurrentWeatherAsync(
+		decimal latitude,
+		decimal longitude,
+		CancellationToken cancellationToken = default) => LegacyCurrentWeatherAsync(new LegacyCurrentWeather.Request {
+			Latitude = latitude,
+			Longitude = longitude,
+			Units = _options.Units
+		}, cancellationToken);
+
+	public async Task<LegacyCurrentWeather.Response> LegacyCurrentWeatherAsync(
+		LegacyCurrentWeather.Request request,
+		CancellationToken cancellationToken = default) {
+		if (cancellationToken.IsSupportedAndCancelled()) {
+			return LegacyCurrentWeather.Response.Cancelled;
+		}
+
+		// ReSharper disable once MethodHasAsyncOverloadWithCancellation
+		var validationResult = _legacyCurrentWeatherValidator.Validate(request);
+
+		if (!validationResult.IsValid) {
+			return LegacyCurrentWeather.Response.Invalid(validationResult);
+		}
+
+		try {
+			var response = await _httpClient.GetAsync(request.GetEndpoint(_options), cancellationToken).ConfigureAwait(false);
+			var currentWeather = await response.Content.ReadFromJsonAsync<LegacyCurrentWeather.Response>(_jsonSerializerOptions, cancellationToken).ConfigureAwait(false);
 
 			currentWeather!.Errors = currentWeather.Error.HasValue()
 				? [currentWeather.Error!]
@@ -59,7 +102,48 @@ internal sealed class OpenWeatherMapClient(
 
 			return currentWeather;
 		} catch {
-			return CurrentWeather.Response.Failed;
+			return LegacyCurrentWeather.Response.Failed;
 		}
 	}
+
+	//public Task<HistoricalWeather.Response> HistoricalWeatherAsync(
+	//	decimal latitude,
+	//	decimal longitude,
+	//	DateTimeOffset startAt,
+	//	DateTimeOffset endAt,
+	//	CancellationToken cancellationToken = default) => HistoricalWeatherAsync(new HistoricalWeather.Request {
+	//		EndAt = endAt,
+	//		Latitude = latitude,
+	//		Longitude = longitude,
+	//		StartAt = startAt,
+	//		Units = _options.Units
+	//	}, cancellationToken);
+
+	//public async Task<HistoricalWeather.Response> HistoricalWeatherAsync(
+	//	HistoricalWeather.Request request,
+	//	CancellationToken cancellationToken = default) {
+	//	if (cancellationToken.IsSupportedAndCancelled()) {
+	//		return HistoricalWeather.Response.Cancelled;
+	//	}
+
+	//	// ReSharper disable once MethodHasAsyncOverloadWithCancellation
+	//	var validationResult = _historicalWeatherValidator.Validate(request);
+
+	//	if (!validationResult.IsValid) {
+	//		return HistoricalWeather.Response.Invalid(validationResult);
+	//	}
+
+	//	try {
+	//		var response = await _httpClient.GetAsync(request.GetEndpoint(_options), cancellationToken).ConfigureAwait(false);
+	//		var historicalWeather = await response.Content.ReadFromJsonAsync<HistoricalWeather.Response>(_jsonSerializerOptions, cancellationToken).ConfigureAwait(false);
+
+	//		historicalWeather!.Errors = historicalWeather.Error.HasValue()
+	//			? [historicalWeather.Error!]
+	//			: historicalWeather.Errors;
+
+	//		return historicalWeather;
+	//	} catch {
+	//		return HistoricalWeather.Response.Failed;
+	//	}
+	//}
 }