+++ Arex388.OpenWeatherMap/OpenWeatherMapClientOptions.cs
@@ -8,4 +8,9 @@ public sealed class OpenWeatherMapClientOptions {
 	/// The API key.
 	/// </summary>
 	public required string Key { get; init; }
+
+	/// <summary>
+	/// The default unit to return results in.
+	/// </summary>
+	public required Units Units { get; init; }
 }